feat: change picture stored type from int to double, fix parser

+++ src/model/picture.cc
@@ -4,11 +4,12 @@
 
 Picture::Picture() : data_(kMaxPictureSize) {}
 
-const std::vector<int>& Picture::GetData() const { return data_; }
+const std::vector<double>& Picture::GetData() const { return data_; }
 
-void Picture::SetData(const std::vector<int>& data) {
+void Picture::SetData(const std::vector<double>& data) {
   if (data.size() != kMaxPictureSize) {
     throw std::invalid_argument("Wrong picture size!");
   }
+
   data_ = data;
 }
